ncvisual_render(): partial renders #170
cuts 'eagle' runtime by about 40%, w00t
