+++ README.md
@@ -1298,12 +1298,11 @@ void ncvisual_destroy(struct ncvisual* ncv);
 // Render the decoded frame to the associated ncplane. The frame will be scaled
 // to the size of the ncplane per the ncscale_e style. A subregion of the
 // frame can be specified using 'begx', 'begy', 'lenx', and 'leny'. To render
-// the entire frame, provide zero for all four arguments. It is an error to
-// otherwise provide zero for either 'lenx' or 'leny', and negative values for
-// any of the four parameters are an error. It is an error to specify any
-// region beyond the boundaries of the frame.
-int ncvisual_render(const struct ncvisual* ncv, int begy, int begx,
-                    int leny, int lenx);
+// the rectangle formed by begy x begx and the lower-right corner, zero can be
+// supplied to 'leny' and 'lenx'. Zero for all four values will thus render the
+// entire visual. Negative values for any of the four parameters are an error.
+// It is an error to specify any region beyond the boundaries of the frame.
+int ncvisual_render(const struct ncvisual* ncv, int begy, int begx, int leny, int lenx);
 
 // Called for each frame rendered from 'ncv'. If anything but 0 is returned,
 // the streaming operation ceases immediately, and that value is propagated out.
