+++ src/demo/eagle.c
@@ -1,5 +1,7 @@
+#include <libavutil/frame.h>
 #include "demo.h"
 
+// display the level map scaled to fit entirely within the visual area
 static struct ncvisual*
 outzoomed_map(struct notcurses* nc, const char* map){
   int averr;
@@ -23,7 +25,7 @@ outzoomed_map(struct notcurses* nc, const char* map){
 static int
 zoom_map(struct notcurses* nc, const char* map){
   int averr;
-  // FIXME determine size that will be represented on screen at once, and how
+  // determine size that will be represented on screen at once, and how
   // large that section has been rendered in the outzoomed map. take the map
   // and begin opening it on larger and larger planes that fit on the screen
   // less and less. eventually, reach our natural NCSCALE_NONE size and begin
@@ -32,61 +34,46 @@ zoom_map(struct notcurses* nc, const char* map){
   if(ncv == NULL){
     return -1;
   }
-  if(ncvisual_decode(ncv, &averr) == NULL){
+  struct AVFrame* frame;
+  if((frame = ncvisual_decode(ncv, &averr)) == NULL){
     ncvisual_destroy(ncv);
     return -1;
   }
+  ncvisual_destroy(ncv);
   // we start at the lower left corner of the outzoomed map
   int truex, truey; // dimensions of true display
   notcurses_term_dim_yx(nc, &truey, &truex);
-  struct ncplane* ncp = ncvisual_plane(ncv);
-  int vx, vy; // dimensions of unzoomed map
-  ncplane_dim_yx(ncp, &vy, &vx);
-  ncplane_move_yx(ncp, truey - vy, truex - vx);
-  if(ncvisual_render(ncv, 0, 0, 0, 0)){
-    ncvisual_destroy(ncv);
-    return -1;
-  }
-  ncplane_move_bottom(ncp);
+  int vwidth = frame->width;
+  int vx = vwidth;
+  int vheight = frame->height; // dimensions of unzoomed map
+  int vy = vheight / 2;
   int zoomy = truey;
   int zoomx = truex;
   while(zoomy < vy && zoomx < vx){
     zoomy += 2;
     zoomx += 2;
-    struct ncplane* zncp = notcurses_newplane(nc, zoomy, zoomx, truey - zoomy, 0, NULL);
-    if(zncp == NULL){
-      ncvisual_destroy(ncv);
-      return -1;
-    }
+    struct ncplane* zncp = notcurses_newplane(nc, zoomy, zoomx, 0, 0, NULL);
     struct ncvisual* zncv = ncplane_visual_open(zncp, map, &averr);
     if(zncv == NULL){
-      ncplane_destroy(zncp);
       ncvisual_destroy(ncv);
       return -1;
     }
     if(ncvisual_decode(zncv, &averr) == NULL){
       ncvisual_destroy(zncv);
-      ncplane_destroy(zncp);
-      ncvisual_destroy(ncv);
       return -1;
     }
-    if(ncvisual_render(zncv, 0, 0, 0, 0)){
+    if(ncvisual_render(zncv, (zoomy - truey) * 2, 0, 0, ((float)truex / zoomx) * zoomx)){
       ncvisual_destroy(zncv);
-      ncplane_destroy(zncp);
-      ncvisual_destroy(ncv);
       return -1;
     }
     if(notcurses_render(nc)){
       ncvisual_destroy(zncv);
-      ncplane_destroy(zncp);
-      ncvisual_destroy(ncv);
       return -1;
     }
     ncvisual_destroy(zncv);
     ncplane_destroy(zncp);
   }
   nanosleep(&demodelay, NULL);
-  ncvisual_destroy(ncv);
   return 0;
 }
