+++ src/lib/notcurses.c
@@ -1605,99 +1605,6 @@ void ncplane_erase(ncplane* n){
   ncplane_unlock(n);
 }
 
-int ncvisual_render(const ncvisual* ncv, int begy, int begx, int leny, int lenx){
-  const AVFrame* f = ncv->oframe;
-  if(f == NULL){
-    return -1;
-  }
-  int x, y;
-  int dimy, dimx;
-  ncplane_dim_yx(ncv->ncp, &dimy, &dimx);
-  ncplane_cursor_move_yx(ncv->ncp, 0, 0);
-  const int linesize = f->linesize[0];
-  const unsigned char* data = f->data[0];
-  for(y = 0 ; y < f->height / 2 && y < dimy ; ++y){
-    if(ncplane_cursor_move_yx(ncv->ncp, y, 0)){
-      return -1;
-    }
-    for(x = 0 ; x < f->width && x < dimx ; ++x){
-      int bpp = av_get_bits_per_pixel(av_pix_fmt_desc_get(f->format));
-      const unsigned char* rgbbase_up = data + (linesize * (y * 2)) + (x * bpp / CHAR_BIT);
-      const unsigned char* rgbbase_down = data + (linesize * (y * 2 + 1)) + (x * bpp / CHAR_BIT);
-/*fprintf(stderr, "[%04d/%04d] %p bpp: %d lsize: %d %02x %02x %02x %02x\n",
-        y, x, rgbbase, bpp, linesize, rgbbase[0], rgbbase[1], rgbbase[2], rgbbase[3]);*/
-      cell c = CELL_TRIVIAL_INITIALIZER;
-      // use the default for the background, as that's the only way it's
-      // effective in that case anyway
-      if(!rgbbase_up[3] || !rgbbase_down[3]){
-        cell_set_bg_alpha(&c, CELL_ALPHA_TRANSPARENT);
-        if(!rgbbase_up[3] && !rgbbase_down[3]){
-          if(cell_load(ncv->ncp, &c, " ") <= 0){
-            return -1;
-          }
-          cell_set_fg_alpha(&c, CELL_ALPHA_TRANSPARENT);
-        }else if(!rgbbase_up[3]){ // down has the color
-          if(cell_load(ncv->ncp, &c, "\u2584") <= 0){ // lower half block
-            return -1;
-          }
-          cell_set_fg_rgb(&c, rgbbase_down[0], rgbbase_down[1], rgbbase_down[2]);
-        }else{ // up has the color
-          if(cell_load(ncv->ncp, &c, "\u2580") <= 0){ // upper half block
-            return -1;
-          }
-          cell_set_fg_rgb(&c, rgbbase_up[0], rgbbase_up[1], rgbbase_up[2]);
-        }
-      }else{
-        cell_set_fg_rgb(&c, rgbbase_up[0], rgbbase_up[1], rgbbase_up[2]);
-        cell_set_bg_rgb(&c, rgbbase_down[0], rgbbase_down[1], rgbbase_down[2]);
-        if(cell_load(ncv->ncp, &c, "\u2580") <= 0){ // upper half block
-          return -1;
-        }
-      }
-      if(ncplane_putc(ncv->ncp, &c) <= 0){
-        cell_release(ncv->ncp, &c);
-        return -1;
-      }
-      cell_release(ncv->ncp, &c);
-    }
-  }
-  flash_damage_map(ncv->ncp->damage, y, true);
-  return 0;
-}
-
-int ncvisual_stream(notcurses* nc, ncvisual* ncv, int* averr, streamcb streamer){
-  ncplane* n = ncv->ncp;
-  int frame = 1;
-  AVFrame* avf;
-  struct timespec start;
-  // FIXME should keep a start time and cumulative time; this will push things
-  // out on a loaded machine
-  while(clock_gettime(CLOCK_MONOTONIC, &start),
-        (avf = ncvisual_decode(ncv, averr)) ){
-    ncplane_cursor_move_yx(n, 0, 0);
-    if(ncvisual_render(ncv, 0, 0, 0, 0)){
-      return -1;
-    }
-    if(streamer){
-      int r = streamer(nc, ncv);
-      if(r){
-        return r;
-      }
-    }
-    ++frame;
-    uint64_t ns = avf->pkt_duration * 1000000;
-    struct timespec interval = {
-      .tv_sec = start.tv_sec + (long)(ns / 1000000000),
-      .tv_nsec = start.tv_nsec + (long)(ns % 1000000000),
-    };
-    clock_nanosleep(CLOCK_MONOTONIC, TIMER_ABSTIME, &interval, NULL);
-  }
-  if(*averr == AVERROR_EOF){
-    return 0;
-  }
-  return -1;
-}
-
 // if "retain_cursor" was set, we don't have these definitions FIXME
 void notcurses_cursor_enable(notcurses* nc){
   if(nc->cnorm){
